+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# Version 0.14.0 (2023-01-30)
+
+## API
+
+- Add a short name for Tool Command Language Script (Tcl Script)
+- Add a short name for UNIX archiver (archiver)
+- Add a short name for UNIX compress (compress)
+
+## New formats support
+
+- Apache Avro Object Container (Avro)
+- Apache Parquet (Parquet)
+- AutoCAD Drawing (DWG)
+- BitTorrent File (Torrent)
+- JPEG-LS (JLS)
+- MP3 URL (M3U)
+- Personal Storage Table (PST)
+
 # Version 0.13.0 (2023-01-22)
 
 ## API
@@ -69,7 +87,7 @@
 - Scalable Vector Graphics (SVG)
 - Shell Script
 - Simple Object Access Protocol (SOAP)
-- Tool Command Language Script
+- Tool Command Language Script (Tcl Script)
 - vCalendar (VCS)
 - vCard (VCF)
 
@@ -126,7 +144,7 @@
 
 - Add MPEG-1/2 Audio Layer 3 (MP3) signature
 - Add Matroska Video (MKV) signature
-- Change UNIX archiver extension from `ar` to `a` (preferred)
+- Change UNIX archiver (archiver) extension from `ar` to `a` (preferred)
 - Fix Apple QuickTime (MOV) signature
 - Fix Joint Photographic Experts Group (JPEG) signature
 
@@ -369,7 +387,7 @@
 - Rich Text Format (RTF)
 - Shapefile (SHP)
 - SketchUp (SKP)
-- UNIX archiver
+- UNIX archiver (archiver)
 
 # Version 0.1.0 (2021-10-03)